+++ CHANGES.rst
@@ -40,6 +40,9 @@ Changelog
   [gbastien]
 - Adapted code to be able to use `collective.ckeditor 4.11+`.
   [gbastien]
+- Make sure annex are not kept upon duplication (to same or distant MC)
+  if annex type requires a PDF file and the annex file is not PDF.
+  [gbastien]
 
 4.2 (2023-03-06)
 ----------------